Maxime/link category

+++ src/com/lat/beans/Reporting.java
@@ -1,18 +1,15 @@
 package com.lat.beans;
 
-
 public class Reporting {
-
-    private Long id;
+    private Integer id;
     private String reason;
     private String moreInformation;
 
-
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
